+++ src/app/beer_garden/api/stomp/processors.py
@@ -1,9 +1,18 @@
 from brewtils.schema_parser import SchemaParser
+from brewtils.models import Operation
 
 
-# Modify to inject custom headers
-def append_headers(response_headers, request_headers=None):
-    return response_headers
+# combines headers to be sent with message
+def append_headers(
+    response_headers=None,
+    request_headers=None,
+    garden_headers=None,
+):
+    headers = {}
+    headers.update(request_headers or {})
+    headers.update(garden_headers or {})
+    headers.update(response_headers or {})
+    return headers
 
 
 class EventManager:
@@ -19,6 +28,10 @@ class EventManager:
 # Processes response messages and event messages to send
 def process_send_message(message):
     many = isinstance(message, list)
+    if message.__class__.__name__ == "Event":
+        message = Operation(
+            operation_type="PUBLISH_EVENT", model=message, model_type="Event"
+        )
     model_class = message.__class__.__name__
     if many:
         model_class = message[0].__class__.__name__